Fixes bug 28870: Callouts are run when moving from grid to form with a shortcut
When a dynamic column is modified, a call is done to the FIC to run the callouts. The list of dynamic columns is stored in the form (in the view form and also in the grid edit form), and is loaded from the
 FIC response. For instance, when a record is created (either or grid or form view), a call is done to the FIC in NEW mode, and the dynamicCols list is initialized.

The cause of this issue is that it was possible to open a record in form view in an inconsistent state, having the form's dynamicCols list not set. In this state the callouts are not run. To get to this in
consistent state, the following steps should take place:
- The user creates a record in the grid (a call to the FIC in NEW mode is done, and the dynamicCols list of the grid edit form is initialized)
- Without saving, the user presses Ctrl+F2 to edit the record in form view. No FIC call is done (there is no need for it), so even though the record is shown in the form view, its dynamicCols list is not s
et. The callouts will not run until a call to the FIC is done (for instance by refreshing the record).

To fix this, the list of dynamicCols of the grid edit form is copied to the form view. This is done in the ksAction_EditInForm function, which is the function executed when the user presses Ctrl+F2 to move
 from the grid view to the form view.

The list of dynamic columns returned from the FIC to the grid view is the same as the one returned from the FIC to the form view, so it is not a problem is a dynamic columns is shown in the form view but n
ot in the grid view.

--HG--
extra : rebase_source : d839114431173c6e7c1e1047f37c11abe8a2fdb6