Fixed issue 28867:It isn't possible to create Menu Items from Widget instances
At first it was possible to create a new menu item entry if the selected Widget was a parent Widget,
but it was not possible if the current widget was an instance of a parent Widget.
For example if you create a new entry in the Query/List widget it wold be possible and it would show in
the Menu Item list of any instance of that Widget. But if you choose a concrete widget which has a
superclass of the type Query/List, and create a new menu item, it would be not possible to display it in the widget.

As it was interesting to be able to add a menu item at widget instance level, the logic which handled this has been changed.
Now it takes into account if the instanciated widget has a menu item.

It is only possible to add functionality to those item menus with functions defined in the OBWidget.js

--HG--
extra : rebase_source : 197d3d06f01ba7bb59269e275e6cfc2f88ca84be