Fixes issue 28754: Context info is properly sent for selectors in P&E grids
The problem was that when a selector drop down was open in a P&E window, the context info sent to the datasource did not include any info of the record being edited. This made impossible to include in the selector a validation based on other fields of the P&E window.

To fix this, the OBSelectorItem.prepareDSRequest function has been updated. Now that function detects if the selector belongs to a P&E grid, and if so it adds to the datasource request params the context info of the record being edited.

--HG--
extra : rebase_source : 12cf4f35439479978fb268962e7168884fa4b445