Fixes bug 29097: A menu tree node can be moved with a template in development
When the Openbravo Trees were reimplemented [1], there was a problem reimplementing the logic done when a Menu tree item is moved. Before the reimplementation, it was possible to move a menu tree item if the module it belonged to was in development or if there was any template module in development. The reimplementation did not take into account the templates, so from then on it has not been possible to moved menu tree nodes having a template in development.

To fix this, templates are taken into accout in the MenuTreeOperationManager class.

[1] https://issues.openbravo.com/view.php?id=25906

--HG--
extra : transplant_source : %F16%0C%E4%B3%D1%86%16S%F7%26bb%96%EAt%3E%16%C5/