Fixed bug 29031: Product's selector does not work on financial's transaction tab
The filter expression for product's search selector, was changed in order to exclude the financial account window, from the "salesPriceList" boolean filter, as this window fails if the "isSalesTransaction" condition is included, just like happens for Commission's window.
Now products's selector works, when adding a transaction for a financial account, even if without the price list version active, because the transaction tab has not this field defined.