Fixes bug 29111: Valued Stock report filter properly by Product Category
A new language parameter was added in the issue that has caused the regression, but in wrong order, causing not to be able to filter by Product Category.
Also a left join was being done instead of inner join, so the query was displaying an wrong extra line with a sum up of all not necessary m_transaction records.