Fixed bug 28973: G/L Journal can be posted in a closed period.
Just like happens in other accounting document clases, the method "loadDocumentDetails" in "DocGLJournal" class was changed to exclude the C_Period_ID field value assignment, which is obtained correctly later in the AcctServer superclass, launching an error when trying to post a G/L Journal in a closes period.
Also the value for period passed as parameter in the method "loadDocumentDetails" is wrong calculated due to the period's organization is not taken into account.
With the changed proposed, the correct journal's period is checked, so if it is closed, an error is shown when posting the document.

--HG--
extra : rebase_source : 82e1af8383536c9a0c238ee1fbca794dd241ace7