Fixed bug 29094: Generate Receipt from Invoice generates too many receipts.
Stored procedure called when Generate Receipt from Invoice process is executed, was returning null when catching the exception launched during regeneration of receitp from the same invoice. This code was removed from the stored procedure, so now the exception is raised and the error message is shown successfuly. No wrong relation is generated between the new Receipt and the Invoice.
Another solution would be to relaunch the exception instead of returning null.
Also display's logic for field "Generate Receipt from Invoice" was changed to show the button when the invoice is processed.