Related to 29124: Code Review.
When opening Add Payment window from Match Statement popup, if the totalselected amount is equal to Zero, the amount coming from bank statement line should be added to Actual Amount. In case the Actual Amount is different to Zero, then it should not be added. In this last case it was replacing the actual payment by bank statement line amount, and it is wrong.

--HG--
extra : rebase_source : 66025535bdee4e105036bafd03a89c423bdd119a