Fixed bug 28685: Goods shipment and invoice relationship not seted in some cases
Goods shipment and invoice relationship was not created when creating shipments using manual form: "Create shipments from orders", so like happens when shipment's lines are created from an specific order, now the invoice's line related to the order, is updated with the shipment's line when it is created from the mentioned manual form, only if this field is null in the invoice's table, i.e. if the invoice is not linked to a previos shipment.
Now goods shipment and invoice relationship is created successfuly when launching "Create shipments from orders" form.
M_InOut_Create was modified so all processes that executed this stored procedure are affected with this change.