Related bug 28999: Grant access to buttons is not working fine in some cases.
The issue had three parts to fix. The part that has been fixed is number two:
The problem was to grant access to buttons when "Secured_process" preference
was "Yes". This problem caused a bad behavior in access to buttons in child tabs.

This bad behavior is reproduced when inherit buttons are created in a non-active
child tab.The buttons are created from "actionToolbarButtons" not in
"toolBar.rightMembers". This was a problem because actionToolbarButtons did not
have "readOnlyIf" property. This property makes accessibility to a process work
correctly.

To solve this problem it has been set the property "readOnlyIf" in
"actionToolbarButtons" at the same time that in "toolBar.rightMembers".Thus when
inherit buttons are created in a child tab, the property "readOnlyIf" exists.
This will keep the same behavior on the inherited buttons in the child tabs.