Fixes issue 29224: Return reason filter works properly in Return from/to P&E
The problem was that in this changeset [1] the way the distinct queries (done to populate the fk filter dropdowns) was modified. The query is different, and the way to process its results too. The regressi
on was caused because some HQL tables build their distinct queries manually in their HQL transformers, and those queries should have been updated too.

The query was changed the whole BaseOBObject in the first position of the result array, instead of including just its id.

[1] https://code.openbravo.com/erp/devel/pi/rev/95dd177fbbc7f18d448b3e31cc12b67dbc0cb87d

--HG--
extra : rebase_source : 5d8c7c154204241332e31e265574d354a4ac99c5