Fixed bug 29119: "Use credit" process should consider defined conversion rate.
The currencies conversion rate, defined for payment's date, is now taken into account when the credit left by the payment, is used during the invoice's process.
Now the field exange rate, is well defined in the new payment created, if exists a conversion between the currencies of the invoice's and the payment's financial account, for the payment's date.