Fixes bug 29249: Transactions created in Match Statement are created processed
On Reconciliation Refactor process and new function was added to process transactions in APRM_MatchingUtility class, replacing the old processTransaction in MatchTransaction servlet. In the new method, the way in which the database stored procedure was being invoked was wrong, so It has been changed in order to work as it was doing in the old method.