Fixes issue 29092: Inconsistent data is not shown in the form view
This issue was reproduced in the following scenario:
- A subtab is open in form view
- A change is done in its parent tab. That change (either through an event handler or through a trigger) deletes the record selected in the subtab and creates a new one
- The subtab is automatically refreshed. When the tab is refreshed:
    * The whole grid is reloaded, it now contains all the current records
    * The form view is shown again, and it tries to obtain the updated records from the grid. As the record selected in the form view is not present in the grid, the form view contains the old values

To fix this, when the record that was selected in the form view is no longer contained in the grid after the tab is refreshed, the form view is not shown. This was the grid view will be visible, and it will show the updated info.