Fixed issue 29220: Total gross amount is not updated properly
If the process of loading the master data fails and the websql database is incomplete, the POS stops working correctly.
Specifically, if the Business Partner information is missing, several javascript errors are risen and as a consequence the POS work flow is stopped. This can prevent the gross of the ticket from calculating, although the total amount button remains active, allowing to complete the ticket without gross amount (and therefore without payments).
In order to prevent this from happening, a new security check has been implemented. If the POS is unable to recover Business Partner information, a popup is shown explaining the situation and prompts the user to reload the window. After reloading a full refresh of the data will be done.

--HG--
extra : rebase_source : 76a822ffb19bd3078cd403928de66db0d2d04169