Fixed bug 28999:It's failing access to defined parent buttons in 2th-3th level
The problem was related to the previous changes. It had added a wrong condition
that prevented a good behaviour in all cases. The wrong condition is:
"if (!view.parentView && button.view.actionToolbarButtons.containsProperty('property', button.property)) {"

Now, it have been resolved all problems by removing the part of the condition
that checks if a view is parent or not. There was no sense.