Fixes issue 29294: valueMap is properly initialized in pick and execute windows
The problem was that when a line in a pick and execute grid was edited, the valueMap of its OBFKComboItems was not properly initialized. The initialization is done in the OBPickAndExecuteGrid.processColumnValue function.

Before the Openbravo combos was refactored, the columnValue parameter used to contain the full valueMap of the combo in its entries attribute. This atribute was passed to the OBPickAndExecuteGrid.setValueMap, and the valueMap of the combos was initialized. The problem was that since the combos was refactored the columnValue parameter no longer contained the entries attribute, because the FIC now only returns the selected value. To fix this, in this case the valueMap is built manually based on the column value and identifier. This valueMap is passed to the OBPickAndExecuteGrid.setValueMap function to initialize the field valueMap.