Fixes bug 29252: Payment method in add payment does not display inactive values
The whereClause of fin_payment_method combo in add payment has been changed to not display never payment methods that are inactive.

Also the default values of Add Payment window that are opened from all windows has been changed to not set as default a Payment Method that is inactive, even if it is marked as default in the financial account or if is the default payment method of the Business Partner.

A change has also been done in OB.APRM.AddPayment.paymentMethodMulticurrency javascript function. In Add Payment window, while clearing Payment Method combo manually, the Financial Account was also being cleared. This was causing a problem when opening from Financial Account, because in this case novalues were displayed on both combos any more. It has been fixed to not to clear the Financial Account combo in this case as always is going to be the same, so it is no necessary to clear it.

--HG--
extra : rebase_source : 78baac12ed012727d61733bf4393d031ed92baa2