Fixes issue 29174: Changes the way summary fields work in widgets
Before this issue [1] was fixed, widget summary fields were computed in the front end, even if the grid had not loaded all the records. This behaviour was wrong, and to fix that issue i
t was decided to compute the summary fields in the backend, the same way it is done in standard grids.

The problem is that widget summary fields that used to show results before fixing [1] (remember that the result that were shown were not valid) do no longer work now, as it is not possi
ble to create a HQL to compute them in the server side. There were three types of summary fields that were not working:
  a) Those based on columns that had an alias in the HQL clause
  b) Those based on columns defined by a subquery
  c) Those based on columns defined in the HQL clase with a summary function

Columns defined like a) are now supported. In order to work they must have a proper value in its Where Clause Left Part field. Columns defined like b) and c) are now no longer supported
. If there is a column of these types in a widget, none of  its summary fields will be shown, and a message will be shown in the log.

[1] https://issues.openbravo.com/view.php?id=26992

--HG--
extra : rebase_source : bf4067cf46e5c5719000b77819085732ad656603