Related to bug 29265: Delete obaprmAllRecordsSelectedByUser in OrderInvoice Load
While executing invalidateCache() function to reload OrderInvoice grid it internally calls to 'OB.APRM.AddPayment.deselectAllRecords' overriden method that sets obaprmAllRecordsSelectedByUser property to true. This property is only needed when manually the select/unselect all checkbox is clicked.

This was causing on certain scenarios that methods to update totals were not being called:
* OB.APRM.AddPayment.updateInvOrderTotal(view.theForm, orderInvoice);
* OB.APRM.AddPayment.updateActualExpected(view.theForm);
* OB.APRM.AddPayment.updateDifference(view.theForm);

So, in order to avoid this in OrderInvoiceOnLoad function, if the property 'obaprmAllRecordsSelectedByUser' exists, it will be deleted.