Fixed bug 29162: Add Transaction process does not consider well amounts of 0.00
The "compareTo" method was taken into account instead of "equals", to compare BigDecimal.ZERO value with 0.00 (a BigDecimal value with a scale), inside AddTransactionFilterExpression class.
This change was made because BigDecimal "equals", compares the value and the scale, returning false when comparing 0 with 0.00, while "compareTo" only compares values, so considers amounts 0.00 and 0 as the same values. Now transaction type is correctly set to BP Deposit, when "dramount" field is set to 0.00.

--HG--
extra : rebase_source : 1c4d52bdf9fd4232223237c1bc165b542d05cbf4