Fixed bug 29353: ICS implemented indirect dataset dependencies
When getting the module dependencies we take into account all the modules installed in the instance instead of only the one selected by the user.
We use the UpdateReferenceDataData.selectModules() method, which is used by the Enterprise Module Management process too, to get the right order.
The UpdateReferenceDataData class is now public to be called from InitialClientSetup.java.
Added missed ProcessingDataset message