Fixes bug 29380: Warehouse field of Manage Reservation P&E shows a proper value
The is a bug in that P&E, reported here [1], that makes the FIC return the wrong value for the warehouse field of the Manage Reservation P&E window. Before [2] was fixed, [1] did not have any consequences, as the value returned by the FIC was not set to the row being edited.

To fix this, the values returned by the FIC are only set if the field is editable. [1] will have to be fixed anyway, because even if now it does not crete problems in this flow, it could be causing problems in others.

[1] https://issues.openbravo.com/view.php?id=29381
[2] https://issues.openbravo.com/view.php?id=28727