Fixed issue 29066: Process definition should inherit permissions from window
The problem was that in some cases when a rol had access to a window, he
did not have access to the processes of the window. The problem was when a
proccess was referenced by a OBUISEL_Selector reference.

This problem happens when is controlled access process. In many cases the
windowId is null.

Now, this problem has been solved because windowId will have a value.