Related to issue 29228: Add a build validation to create an alert if wrong data
A build validation will be executed to check if it exists wrong data (g/l journal accounting with wrong amounts) related to the issue. In that case an alert will be created to tell the user to reset accounting, but when fixed it will be deactivated to avoid execute it again.
The build validation will only be executed when updating from Q1 or Q1.1, because the issue is introduced in Q1 and resolved in Q1.2.