Fixed issue 29412:It is not possible to use the new reporting class everywere.
The problem was that calling the methods of the "ExportType" class inside the "ReportingUtils" class
from some parts of the code (the old code) was not working properly.
The problem was that it was needed to have the AdminMode set to true to access to a table which
was accessed in the constructor of the "ExportType" class.
The new code has the AdminMode set to true, that is why it works.
But in the old code it was not.

For fixing this, the AdminMode has been set to true inside the constructor of the "ExportType" class
itself. It is needed for the download of the report.
Other change that has been made is changing the type of the variable filetype, now it is String. This is
for saving memory.
It has been taken into account to the fact of having a different type of report in the future. This has been
handled with the "fileType = "application/" + extension" code. Now new extensions will be supported.
Now whereever one calls those methods, it works properly.