Fixed issue 29427: Modifiying the object of the enum ExportType is allowed.
The problem is that with the method getExportParemeters of the enum class ExportType, it is possible to get those
parameters in the Hasmap and after that, modify the object with the method "put".
This should not be allowed because this object belongs to an enum and all the instances have to share the same object,
we don't want to allow the modification of it.

For fixing this, an instance of the hashmap is done for making sure that if this method is called, it is only accessing
to the parameters of the current Map instance.