Fixed issue 29441: It is not possible to export a report in XLS format.
Trying to export a report in XLS format, using the following module does not work:
 https://code.openbravo.com/erp/mods/org.openbravo.platform.features

It shows a casting error in the log. The problem is in the ReportingUtils class.
When it tries to export a report of XLS type, the variable "target" is type "File".
The problem is when the JRExporterParameter class of Jasper is used:

exportParameters.put(JRExporterParameter.OUTPUT_FILE_NAME, target);

The target needs to be of type "String" instead of "File" for the correct execution of
the code above.

To fix this the method "getAbsolutePath" of the "File" class has been called for getting
the absolute path which is what is needed, but  having the type as "String" instead
of "File".