Fixes bug 29430: Net Unit Price in Costing is properly calculated
Using Cost Adjustments the Net Unit Price of a product on a transaction could be affected with those. Starting from now the Net Unit Price assigned in Costing tab is going to take into account the Cost Adjustments of type Unit Cost.

So the calculation would be: ((originalPrice * qty) + adjustmentsAmount) / qty.