Fixed bug 28702: Assets Organization should not be changed if it has lines
Like happens in other windows such as Products and Business Partner, the organization's field behaviour in Assets window was changed to not be updatable.
Now when the Assets record is saved, it is not possible to change its organization, so there will not be inconsistencies with previous created lines if the amortizations are recalculated.