Related to issue 29302: Changed the way a new ticket is created with multiorders
After closing the last receipt in a multiorder ticket, a new receipt will be created if the receipt list is empty (this is done automatically when deleting an order). At the moment of finishing closing the multiorder we only need to synchronize the open ticket if needed, in order to update its document number.