Fixes bug 29473: Bank Statement Line amount maintained in Add Payment from FA
If exists bslamount this means that Add Payment window has been opened from Match Statement window, so, when selecting a line in Order/Invoices grid, this amount will be set in the line when is smaller than outstanding amount.

If Outstanding amount is smaller, then outstanding amount will be assigned.