Fixed issue 29504: It is possible to create a payment in/out
It is possible to create a payment header when a payment method or a financial
account is disabled in a given financial account.

Now only active payment methods of financial accounts and active financial
accounts are showing in the payment in / payment out window.

In addition, if the user selects a business partner in these windows and the
default payment method and/or the default financial account is/are inactive,
a warning message is shown to the user, indicating which one is inactive.

These warning messages are also included in the fix.