Fixed issue 29496: Not possible to do return to vendor using non stocked product
It is not possible to do a return to vendor using non stocked products.

The Pick and Edit window didn't show order from non stocked produts caused by
a Hibernate error when creating the sql from the HQL Where Clause.

The solution is to change the HQL where clause from this window and avoid the
cross join of the m_locator table. A subselect query has been added to the
Where Clause to avoid this error.

In addition, a new constraint has been added to the HQL Where Clause and now
the products without a storage bin are shown as well.