Fixes bug 29542: Payment Out can be reactivated without bp purchase pricelist
The problem is that the method 'getConversionRateDocumentForInvoice' tries to get a conversion using the currency of Purchase PriceList of the business partner, and it could happen to be empty, so it fails.

Starting from 3.0PR14Q4 the Currency field in the business partner is mandatory, so it can be used for Sales/Purchase flow.

It could be that at this point while reactivating the Payment to be the Currency of the Business Partner null, because it was created long time ago and no currency set yet, so, if this is the case an error will be raised to the user in order to fix this.