Fixes bug 29448, Fixes bug 29447, Fixes bug 29449
The following changes was made in Stock Reservation pick and edit:
When filters are executed, a Reservation entity is created from the json parameter @MaterialMgmtReservation.id@. This works fine when the pop-up is called from Stock Reservation window, but from Sales Order, this parameter does not exists, and the filter process fails. Now is included a code line that puts a @MaterialMgmtReservation.id@ into the json parameters before the filters are applied.
Array critera filtering was bad created when getting grid's data, so was changed, to allow filter's multiple selection.
An hibernate exception is launched when typing wrong filters and the expression 'in ()' is included in the query to be executed. So this expression was replaced by one which returns a false value: '1 = 2'
Were included also, changes to allow filtering quantity's column. Other changes were included to avoid exceptions comparing to null values.