Fixes bug 29541: Payment Out can be reactivated without bp purchase pricelist
The problem is that the method 'getConversionRateDocumentForInvoice' tries to get a conversion using the currency of Purchase PriceList of the business partner, and it could happen to be empty, so it fails.

Starting from 3.0PR14Q4 the Currency field in the business partner is mandatory, so it can be used for Sales/Purchase flow.

It could be that at this point while reactivating the Payment to be the Currency of the Business Partner null, because it was created long time ago and no currency set yet, so, if this is the case an error will be raised to the user in order to fix this.

A last change has been made in 'OBMessageUtils.messageBD' method, adding an extra parameter when it is wanted to avoid doing escape of '\n' and doublequotes characters. In this case it is needed because the message that needs to be displayed needs to maintain them because contains a link. This implementation is similar to the one found in BasicUtility.messageBD.