Fixed issue 29574: It is not possible to reactivate a Landed Cost
It is not possible to reactivate a Landed Cost under some circunstances.
The problem was fixed from the LandedCostDistributionAlgorithm.java

When you try to reactivate a Landed Cost, an error message appears and
the process was cancelled

Now the process is done succesfully and the Landed Cost can be reactivated.

--HG--
extra : rebase_source : c16aaa9c0dfa444816419f8d7ee7217a8f2f5355