Fixed issue 29584: It is not possible to reactivate a Landed Cost
It is not possible to reactivate a Landed Cost under some circunstances.
The problem was fixed from the LandedCostDistributionAlgorithm.java

When you try to reactivate a Landed Cost, an error message appears and
the process was cancelled

Now the process is done succesfully and the Landed Cost can be reactivated.