fixes issue 29560: added function to check if the data is loaded in p&e grids
A check is added to prevent the execution of the on grid load functions if the local data is not loaded.
This problem could happen when a request is done and the load of a previous request has not finished.

--HG--
extra : rebase_source : 19e9ace62ab5448c4dcf0bd7e144937424a2ccd3