Fixed issue 29226: Value stock report is slow with many rows on m_transaction
The value stock reports is slow with many rows on the m_transaction_cost table
and most of the them are for the same locator.

The problem was in the sql query from the report. This query needed an
improvement.

The sql query from ReportValuationStockData has changed and now the report is
faster and the return values are the same as before.

--HG--
extra : rebase_source : 82ddefe5ca9d33c6c99e5b17dfa3b57e883e0fa8