Fixed issue 29521: Storage Bin and UOM fields are not filled well
Storage Bin and UOM fields are not filled when posting Work Efforts and
Bill of Material Productions.

The problem was in the Production side of the ERP for both posts.

The solution is to retrieve the UOM from the M_PRODUCTIONLINE table, and
set the storage bin and the UOM in the fact lines.