Fixed bug 29607: setup application is not taking into account the context name
The setup application is not taking into account the context name that users
are configuring. The values of "deploy-name" and "context-root" properties
of org.eclipse.wst.common.component are not updating.

Now it has been solved and if user set "context-name" in setup application.
this value is updating in org.eclipse.wst.common.component. For example if
a user set the context name as "test" the URL to go to Openbravo is
"http://localhost:8080/test" instead of default one "http://localhost:8080/openbravo".