Fixed issue 29531: Not possible to create standards with a discarded toolset
Not possible to create standards if a toolset is defined as Discard.

The problem occurs when, in the ma_productionrun_standard function, insert
a null ma_toolset_id. This happens when only discard toolsets are available.

The solution is to avoid the insert of a null ma_toolset_id  in the function.

Now,  it is possible to create a Work Effort with a discarded toolset. In the
Work Effort window, the discarded toolset doesn't appear in the Toolset tab.

--HG--
extra : rebase_source : 12edc94809955fe4111728d2074ce5199a14f5b7