Fixed issue 29571: Book Quantity is not updated to 0 by Update Quantity process
In Physical Inventory Lines, Book Quantity is not updated to 0 by Update
Quantity process if product has a mandatory attribute.

The decided solution is to delete the lines from the physical inventory in two
cases: if the product of the line doesn't had an associated record in the
m_storage_detail table, or if the quantity on hand is equal to 0 or the order
quantity is equal to 0 or null.

--HG--
extra : rebase_source : 88f157a10567d916ef586de41d4f1ee8d12d9602