fixes issue 29554: Date filters not cleared when switching between saved views
The problem was that the clearValue() method of FormItem is not enough to clear the contents of the filter.
To solve it the clearValue method of OBMiniDateRangeItem has been overriden.
Also, it was necessary to call setValue() inside the displayValue() method, to set the value of the object and avoid the clearing of the dates in the filter when smartclient invokes the setItemValues() method of the form.

--HG--
extra : rebase_source : 61b429f13efb6634f133d4cb7da61e65bdce2d0d