Fixed issue 29064:There are problems with the organization of the notes.
There were two problems:
First: When a note was created, it had context organization instead of document organization.
To fix this, a new field which stores the note's organization has been created in the "ob-view-form-notes.js".
Second: If a note was added with an organization and after that the organization of the note
was changed, there were problems to handle it.
A new manual datasource has been created for handling the fetch,add and remove operations of the notes,
making sure that the organization filter is disabled. So, we grant the fact that if a user of one
organization creates a note and after that the organization of the note is changed to another one which
the user has not access, the user will be able to see or delete the note.